refactor(uniffi): extract fire-uniffi-topics
Phase 3d of the uniffi multi-namespace split (largest domain).

Rust:
- Add rust/crates/fire-uniffi-topics with
  setup_scaffolding!("fire_uniffi_topics"). Hosts FireTopicsHandle plus the
  topic list/detail/post/poll/upload/vote record types. The 21 topic-domain
  methods move off FireAppCore verbatim: fetch_topic_list, fetch_topic_detail,
  fetch_topic_detail_initial, fetch_topic_posts, create_reply, fetch_post,
  update_post, create_topic, create_private_message, update_topic,
  upload_image, lookup_upload_urls, report_topic_timings, like_post,
  unlike_post, toggle_post_reaction, vote_poll, unvote_poll, vote_topic,
  unvote_topic, fetch_topic_voters.
- FireAppCore gains topics: Arc<FireTopicsHandle> + topics() getter.
  state_topic_detail.rs and state_topic_list.rs are deleted. VoteResponseState
  and VotedUserState (miscategorised in state_user.rs) move with topics since
  they are part of the vote API surface.
- TopicListQueryState moves with topics rather than types because it is a
  topics-domain input only (types already holds TopicListKindState and the
  shared TopicListState response shape that notifications/topics both consume).
- Regression test for TopicDetailState::interaction_count moves from
  fire-uniffi's lib.rs to fire-uniffi-topics/src/lib.rs alongside its
  From<TopicDetail> impl.

iOS / Android:
- FireSessionStore on both platforms routes topic/list/post/reply/upload/
  reaction/poll/vote calls through core.topics(). Kotlin imports for
  TopicDetailQueryState / TopicDetailState / TopicListQueryState /
  TopicPostState move to uniffi.fire_uniffi_topics; TopicListKindState moves
  to uniffi.fire_uniffi_types (it already moved with other shared list types
  in an earlier phase). MainActivity.kt and TopicDetailActivity.kt picked up
  the import renames.

Build plumbing:
- native/ios-app/project.yml preBuildScript outputs gain
  fire_uniffi_topics.swift. Regenerated project.pbxproj.

cargo test --workspace stays green.
