+++ native/ios-app/project.yml
@@ -112,6 +112,7 @@ targets:
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_messagebus.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_notifications.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_search.swift
+          - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_topics.swift
           - $(SRCROOT)/Generated/FireUniFfi/fire_uniffi_types.swift
           - $(SRCROOT)/Generated/fire_uniffiFFI/fire_uniffiFFI.h
           - $(SRCROOT)/Generated/fire_uniffiFFI/module.modulemap
