+++ rust/crates/fire-uniffi-topics/src/records.rs
@@ -1,13 +1,64 @@
 use fire_models::{
     InviteCreateRequest, Poll, PollOption, PostReactionUpdate, PostUpdateRequest,
     PrivateMessageCreateRequest, ResolvedUploadUrl, TopicCreateRequest, TopicDetail,
-    TopicDetailCreatedBy, TopicDetailMeta, TopicDetailQuery, TopicPost, TopicPostStream,
-    TopicReaction, TopicReplyRequest, TopicThread, TopicThreadFlatPost, TopicThreadReply,
-    TopicThreadSection, TopicTimelineEntry, TopicTimingEntry, TopicTimingsRequest,
-    TopicUpdateRequest, UploadResult,
+    TopicDetailCreatedBy, TopicDetailMeta, TopicDetailQuery, TopicListQuery, TopicPost,
+    TopicPostStream, TopicReaction, TopicReplyRequest, TopicThread, TopicThreadFlatPost,
+    TopicThreadReply, TopicThreadSection, TopicTimelineEntry, TopicTimingEntry,
+    TopicTimingsRequest, TopicUpdateRequest, UploadResult, VoteResponse, VotedUser,
 };
 
-use fire_uniffi_types::{TopicParticipantState, TopicTagState};
+use fire_uniffi_types::{TopicListKindState, TopicParticipantState, TopicTagState};
+
+#[derive(uniffi::Record, Debug, Clone)]
+pub struct TopicListQueryState {
+    pub kind: TopicListKindState,
+    pub page: Option<u32>,
+    pub topic_ids: Vec<u64>,
+    pub order: Option<String>,
+    pub ascending: Option<bool>,
+    pub category_slug: Option<String>,
+    pub category_id: Option<u64>,
+    pub parent_category_slug: Option<String>,
+    pub tag: Option<String>,
+    pub additional_tags: Vec<String>,
+    pub match_all_tags: bool,
+}
+
+impl From<TopicListQuery> for TopicListQueryState {
+    fn from(value: TopicListQuery) -> Self {
+        Self {
+            kind: value.kind.into(),
+            page: value.page,
+            topic_ids: value.topic_ids,
+            order: value.order,
+            ascending: value.ascending,
+            category_slug: value.category_slug,
+            category_id: value.category_id,
+            parent_category_slug: value.parent_category_slug,
+            tag: value.tag,
+            additional_tags: value.additional_tags,
+            match_all_tags: value.match_all_tags,
+        }
+    }
+}
+
+impl From<TopicListQueryState> for TopicListQuery {
+    fn from(value: TopicListQueryState) -> Self {
+        Self {
+            kind: value.kind.into(),
+            page: value.page,
+            topic_ids: value.topic_ids,
+            order: value.order,
+            ascending: value.ascending,
+            category_slug: value.category_slug,
+            category_id: value.category_id,
+            parent_category_slug: value.parent_category_slug,
+            tag: value.tag,
+            additional_tags: value.additional_tags,
+            match_all_tags: value.match_all_tags,
+        }
+    }
+}
 
 #[derive(uniffi::Record, Debug, Clone)]
 pub struct TopicDetailQueryState {
@@ -581,6 +632,48 @@ impl From<TopicDetail> for TopicDetailState {
             flat_posts: value.flat_posts.into_iter().map(Into::into).collect(),
             timeline_entries: value.timeline_entries.into_iter().map(Into::into).collect(),
             details: value.details.into(),
+        }
+    }
+}
+
+#[derive(uniffi::Record, Debug, Clone)]
+pub struct VotedUserState {
+    pub id: u64,
+    pub username: String,
+    pub name: Option<String>,
+    pub avatar_template: Option<String>,
+}
+
+impl From<VotedUser> for VotedUserState {
+    fn from(value: VotedUser) -> Self {
+        Self {
+            id: value.id,
+            username: value.username,
+            name: value.name,
+            avatar_template: value.avatar_template,
+        }
+    }
+}
+
+#[derive(uniffi::Record, Debug, Clone)]
+pub struct VoteResponseState {
+    pub can_vote: bool,
+    pub vote_limit: u32,
+    pub vote_count: i32,
+    pub votes_left: i32,
+    pub alert: bool,
+    pub who_voted: Vec<VotedUserState>,
+}
+
+impl From<VoteResponse> for VoteResponseState {
+    fn from(value: VoteResponse) -> Self {
+        Self {
+            can_vote: value.can_vote,
+            vote_limit: value.vote_limit,
+            vote_count: value.vote_count,
+            votes_left: value.votes_left,
+            alert: value.alert,
+            who_voted: value.who_voted.into_iter().map(Into::into).collect(),
         }
     }
 }
