--- rust/crates/fire-uniffi/src/state_topic_list.rs
@@ -1,54 +1,0 @@
-use fire_models::TopicListQuery;
-
-use fire_uniffi_types::TopicListKindState;
-
-#[derive(uniffi::Record, Debug, Clone)]
-pub struct TopicListQueryState {
-    pub kind: TopicListKindState,
-    pub page: Option<u32>,
-    pub topic_ids: Vec<u64>,
-    pub order: Option<String>,
-    pub ascending: Option<bool>,
-    pub category_slug: Option<String>,
-    pub category_id: Option<u64>,
-    pub parent_category_slug: Option<String>,
-    pub tag: Option<String>,
-    pub additional_tags: Vec<String>,
-    pub match_all_tags: bool,
-}
-
-impl From<TopicListQuery> for TopicListQueryState {
-    fn from(value: TopicListQuery) -> Self {
-        Self {
-            kind: value.kind.into(),
-            page: value.page,
-            topic_ids: value.topic_ids,
-            order: value.order,
-            ascending: value.ascending,
-            category_slug: value.category_slug,
-            category_id: value.category_id,
-            parent_category_slug: value.parent_category_slug,
-            tag: value.tag,
-            additional_tags: value.additional_tags,
-            match_all_tags: value.match_all_tags,
-        }
-    }
-}
-
-impl From<TopicListQueryState> for TopicListQuery {
-    fn from(value: TopicListQueryState) -> Self {
-        Self {
-            kind: value.kind.into(),
-            page: value.page,
-            topic_ids: value.topic_ids,
-            order: value.order,
-            ascending: value.ascending,
-            category_slug: value.category_slug,
-            category_id: value.category_id,
-            parent_category_slug: value.parent_category_slug,
-            tag: value.tag,
-            additional_tags: value.additional_tags,
-            match_all_tags: value.match_all_tags,
-        }
-    }
-}
